+++ R/wb_lmer.R
@@ -447,9 +447,9 @@ summary.wbm <- function(object, ...) {
                         model = est_name)
 
   coefs <- j$coeftable
-  rownames(coefs) <- gsub("`", "", rownames(coefs), fixed = TRUE)
+  rownames(coefs) <- un_bt(rownames(coefs))
   if (!is.null(x2$ints)) {
-    x2$ints <- gsub("`", "", x2$ints, fixed = TRUE)
+    x2$ints <- un_bt(x2$ints)
   }
 
   varying <- x2$varying